requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
python-dotenv==1.1.1
whisper==1.1.10
tqdm==4.67.1
torch==2.7.1
click==8.2.1
typing_extensions==4.14.1
